+++ arch/Makefile.am
@@ -1,4 +1,4 @@
-AM_CPPFLAGS += -I$(top_srcdir)/ccutil
+AM_CPPFLAGS += -I$(top_srcdir)/ccutil -I$(top_srcdir)/viewer
 AUTOMAKE_OPTIONS = subdir-objects
 SUBDIRS =
 AM_CXXFLAGS =
@@ -8,15 +8,17 @@ AM_CXXFLAGS += -fvisibility=hidden -fvisibility-inlines-hidden
 AM_CPPFLAGS += -DTESS_EXPORTS
 endif
 
-include_HEADERS = \
-	dotproductavx.h dotproductsse.h
+include_HEADERS = dotproductavx.h dotproductsse.h simddetect.h
 
-noinst_HEADERS = 
+noinst_HEADERS =
 
 if !USING_MULTIPLELIBS
 noinst_LTLIBRARIES = libtesseract_avx.la libtesseract_sse.la
+noinst_LTLIBRARIES += libtesseract_arch.la
 else
 lib_LTLIBRARIES = libtesseract_avx.la libtesseract_sse.la
+lib_LTLIBRARIES += libtesseract_arch.la
+libtesseract_arch_la_LDFLAGS = -version-info $(GENERIC_LIBRARY_VERSION)
 libtesseract_avx_la_LDFLAGS = -version-info $(GENERIC_LIBRARY_VERSION)
 libtesseract_sse_la_LDFLAGS = -version-info $(GENERIC_LIBRARY_VERSION)
 endif
@@ -27,6 +29,8 @@ endif
 if SSE41_OPT
 libtesseract_sse_la_CXXFLAGS = -msse4.1
 endif
+
+libtesseract_arch_la_SOURCES = simddetect.cpp
 
 libtesseract_avx_la_SOURCES = dotproductavx.cpp
