+++ lstm/weightmatrix.cpp
@@ -18,88 +18,13 @@
 
 #include "weightmatrix.h"
 
-#undef X86_BUILD
-#if defined(__x86_64__) || defined(__i386__) || defined(_WIN32)
-# if !defined(ANDROID_BUILD)
-#  define X86_BUILD 1
-# endif // !ANDROID_BUILD
-#endif // x86 target
-
-#if defined(X86_BUILD)
-# if defined(__linux__) || defined(__MINGW32__)
-#  include <cpuid.h>
-# elif defined(_WIN32)
-#  include <intrin.h>
-# endif
-#endif
 #include "dotproductavx.h"
 #include "dotproductsse.h"
+#include "simddetect.h"
 #include "statistc.h"
-#include "svutil.h"
 #include "tprintf.h"
 
 namespace tesseract {
-
-// Architecture detector. Add code here to detect any other architectures for
-// SIMD-based faster dot product functions. Intended to be a single static
-// object, but it does no real harm to have more than one.
-class SIMDDetect {
- public:
-  SIMDDetect()
-      : arch_tested_(false), avx_available_(false), sse_available_(false) {}
-
-  // Returns true if AVX is available on this system.
-  bool IsAVXAvailable() {
-    if (!arch_tested_) TestArchitecture();
-    return avx_available_;
-  }
-  // Returns true if SSE4.1 is available on this system.
-  bool IsSSEAvailable() {
-    if (!arch_tested_) TestArchitecture();
-    return sse_available_;
-  }
-
- private:
-  // Tests the architecture in a system-dependent way to detect AVX, SSE and
-  // any other available SIMD equipment.
-  void TestArchitecture() {
-    SVAutoLock lock(&arch_mutex_);
-    if (!arch_tested_) {
-#if defined(X86_BUILD)
-# if defined(__linux__) || defined(__MINGW32__)
-      unsigned int eax, ebx, ecx, edx;
-      if (__get_cpuid(1, &eax, &ebx, &ecx, &edx) != 0) {
-        sse_available_ = (ecx & 0x00080000) != 0;
-        avx_available_ = (ecx & 0x10000000) != 0;
-      }
-# elif defined(_WIN32)
-      int cpuInfo[4];
-      __cpuid(cpuInfo, 0);
-      if (cpuInfo[0] >= 1) {
-        __cpuid(cpuInfo, 1);
-        sse_available_ = (cpuInfo[2] & 0x00080000) != 0;
-        avx_available_ = (cpuInfo[2] & 0x10000000) != 0;
-      }
-# endif
-      if (avx_available_) tprintf("Found AVX\n");
-      if (sse_available_) tprintf("Found SSE\n");
-#endif // X86_BUILD
-      arch_tested_ = true;
-    }
-  }
-
- private:
-  // Detect architecture in only a single thread.
-  SVMutex arch_mutex_;
-  // Flag set to true after TestArchitecture has been called.
-  bool arch_tested_;
-  // If true, then AVX has been detected.
-  bool avx_available_;
-  // If true, then SSe4.1 has been detected.
-  bool sse_available_;
-};
-
-static SIMDDetect detector;
 
 // Copies the whole input transposed, converted to double, into *this.
 void TransposedArray::Transpose(const GENERIC_2D_ARRAY<double>& input) {
@@ -258,7 +183,7 @@ void WeightMatrix::MatrixDotVector(const inT8* u, double* v) const {
   for (int i = 0; i < num_out; ++i) {
     const inT8* Wi = wi_[i];
     int total = 0;
-    if (detector.IsSSEAvailable()) {
+    if (SIMDDetect::IsSSEAvailable()) {
       total = IntDotProductSSE(u, Wi, num_in);
     } else {
       for (int j = 0; j < num_in; ++j) total += Wi[j] * u[j];
@@ -410,8 +335,8 @@ double WeightMatrix::DotProduct(const double* u, const double* v, int n) {
   // is about 8% faster than sse. This suggests that the time is memory
   // bandwidth constrained and could benefit from holding the reused vector
   // in AVX registers.
-  if (detector.IsAVXAvailable()) return DotProductAVX(u, v, n);
-  if (detector.IsSSEAvailable()) return DotProductSSE(u, v, n);
+  if (SIMDDetect::IsAVXAvailable()) return DotProductAVX(u, v, n);
+  if (SIMDDetect::IsSSEAvailable()) return DotProductSSE(u, v, n);
   double total = 0.0;
   for (int k = 0; k < n; ++k) total += u[k] * v[k];
   return total;
